app/requirements.txt
streamlit
fastapi
uvicorn
requests
pydantic >= 2.0.0

google-genai        
gradio-client           
google-api-python-client 

matplotlib
tqdm
python-dotenv
tenacity       

torch
transformers
peft
bitsandbytes
accelerate
sentence-transformers
python-dotenv